pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-studio"
version = "0.1.0"
description = "AI Studio local training platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "jinja2>=3.1",
  "pillow>=10.0",
  "pytorch-lightning>=2.2",
  "torch>=2.2",
  "torchmetrics>=1.3",
  "faster-coco-eval>=1.6",
  "torchvision>=0.17",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "scikit-learn>=1.4",
  "onnx>=1.15",
  "onnxruntime>=1.17",
  "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "UP"]
